pyproject.toml
[project]
name = "gdm-faker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "clickhouse-connect>=0.8.6",
    "faker>=29.0.0",
    "pandas>=2.2.3",
    "polars>=1.9.0",
    "prefect>=3.1.0",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.0.1",
]
